restart mgmt over-usb interface if not RUNNING
PROBLEM:
Sometimes after "config reload" command the management
interface (eth0) connected over ETH-USB device is
  UP/RUNNING on a remote peer side, but
  stays UP/not-running on the local side.
Well-known negotiation problem of ETH-over-USB devices (ASIX).

SOLUTION:
If the management interface is a USB device and is not RUNNING
(/sys/class/net/${if}/operstate is not "up") at the end of the
"config reload" command
bring it down and up with delay to trigger renegotiation.

Signed-off-by: Yan Markman <ymarkman@marvell.com>